+++ helpers/settings.py
@@ -14,7 +14,6 @@ extensions = [
     'cogs.stats',
     'cogs.weather',
     'cogs.swear',
-    'cogs.overwatch',
     'cogs.versioning',
     'cogs.games',
 ]
@@ -36,6 +35,5 @@ channels = {
     "tabletop": "188081332834926592",
     "technology": "188085243478147074",
     "programming": "188083195470807040",
-    # "giveaways": "181233418099490826",  # for the testing server
     "giveaways": "188087131070005248",
 }