+++ cogs/versioning/changelog.md
@@ -1,13 +1,25 @@
 Changelog
 =========
 
-1.0
-===
+1.0.2
+=====
+
+* Added support for community games
+    * This system is a prototype, if you would like to see any new functionality or spot a bug please use our issue tracker
+     https://github.com/iScrE4m/IdiotechDiscordBot/issues
+
+* [!game][]                             list opened community games
+* [!game][open <gamename>]                         open a game for everyone to join you
+* [!game][description <description>]    provide everyone who joins with information about your game
+* [!game][join <gamename>]              join an opened community game
+* [!game][leave]                        leave the game you joined
+* [!game][close]                        close your community game (please don't forget)
+* For more detailed information you can always use [!help][command (subcommand)]
+_____
 
 1.0.1.2
 -------
-* Moved about some code
-* Hello
+* Preparation for [!games][] command
 
 1.0.1.1
 -------
@@ -16,7 +28,7 @@ Changelog
 * [!release][] Added Mighty No. 9 and WoW: Legion to list
 
 1.0.1
------
+=====
 * Added Versioning system
 * Everything up to this point is only found in commits
 * [!version][] says bot's version out loud